+++ src/refinegems/curation/db_access/kegg.py
@@ -415,81 +415,3 @@ def kegg_gene_comp(model: libModel, organismid: str, gff_file: str) -> pd.DataFr
     return missing_reactions
 
 
-
-# @TODO open issues
-def kegg_reaction_parser(rn_id:str) -> dict: 
-    """Get the entry of a KEGG reaction ID and 
-    parse the information into a dictionary.
-
-    Args:
-        - rn_id (str): 
-            A reaction ID existing in KEGG.
-
-    Returns:
-        dict: 
-            The KEGG entry information as a dictionary.
-    """
-
-    # get KEGG reaction entry
-    try:
-        kegg_reac = REST.kegg_get(F'rn:{rn_id}')
-        kegg_reac = kegg_reac.read()
-    except Exception as e:
-        # @TODO
-        return None
-
-    # parse the entry for necessary information
-    features = {}
-    db_entries = []
-    pathways = []
-    rc = []
-    references = {'kegg.reaction':rn_id}
-    collect = False
-    for line in kegg_reac.split('\n'):
-        if line:
-            if line.startswith('NAME'):
-                features['name'] = line.replace('NAME','',1).strip()
-            elif line.startswith('EQUATION'):
-                features['equation'] = line.replace('EQUATION','',1).strip()
-            elif line.startswith('ENZYME'):
-                references['ec-code'] = line.replace('ENZYME','',1).strip()
-            elif line.startswith('RCLASS'):
-                rc.append(line.replace('RCLASS','',1).strip().split(' ')[0])
-                collect = True
-            elif line.startswith('PATHWAY'):
-                pathways.append(line.replace('PATHWAY','',1).strip().split(' ')[0])
-                collect = True
-            elif line.startswith('DBLINKS'):
-                db_entries.append(line.replace('DBLINKS','',1).strip())
-                collect = True
-            elif collect == True and line[0] != '/':
-                if len(db_entries) == 0:
-                    if line[0].isupper():
-                        collect = False
-                    else:
-                        line = line.strip()
-                        if line.startswith('RC'):
-                            rc.append(line.split(' ')[0])
-                        else:
-                            pathways.append(line.split(' ')[0])
-                else:
-                    db_entries.append(line.strip())
-            else:
-                continue
-
-    # parse references
-    for entry in db_entries:
-        db, identifier = entry.split(':')
-        db = db.strip().lower()
-        if db in references:
-            references[db] = references[db].append(identifier)
-        else:
-            references[db] = [identifier.strip()]
-    if len(pathways) > 0:
-        references['kegg.pathway'] = pathways
-    if len(rc) > 0:
-        references['kegg.rclass'] = rc
-    if len(references) > 0:
-        features['db'] = references
-
-    return features
