+++ src/refinegems/utility/entities.py
@@ -26,12 +26,8 @@ from string import ascii_uppercase, digits
 from typing import Union, Literal
 
 from .cvterms import add_cv_term_genes, add_cv_term_metabolites, add_cv_term_reactions
-from .io import search_ncbi_for_gpr, load_a_table_from_database
+from .io import search_ncbi_for_gpr, load_a_table_from_database, kegg_reaction_parser
 from ..developement.decorators import *
-# @TODO @TEST for import problems 
-# # -> afraid, this might cause a circular import (due to curation)
-from ..curation.db_access.db import parse_reac_str
-from ..curation.db_access.kegg import kegg_reaction_parser
 
 ################################################################################
 # variables
@@ -925,6 +921,108 @@ def build_metabolite_biocyc(id:str, model:cobra.Model,
 
 # adding reactions cobra models
 # -----------------------------
+
+# @TODO: BioCyc missing
+def parse_reac_str(equation:str, 
+                   type:Literal['BiGG','BioCyc','MetaNetX','KEGG']='MetaNetX') -> tuple[dict,dict,list,bool]:
+    """Parse a reaction string.
+
+    Args:
+        - equation (str): 
+            The equation of a reaction as a string (as saved in the database).
+        - type (Literal['BiGG','BioCyc','MetaNetX','KEGG'], optional): 
+            The name of the database the equation was taken from. 
+            Can be 'BiGG','BioCyc','MetaNetX','KEGG'.
+            Defaults to 'MetaNetX'.
+
+    Returns:
+        tuple: 
+            
+            Tuple of (1) dict, (2) dict, (3) list & (4) bool:
+            
+            1. Dictionary with the reactant IDs and their stoichiometric factors.
+            2. Dictionary with the product IDs and their stoichiometric factors.
+            3. List of compartment IDs or None, if they cannot be extract from the equation.
+            4. True, if the reaction is reversible, else False.
+    """
+
+    products = {}
+    reactants = {}
+    compartments = list()
+    is_product = False
+    reversible = True
+
+    match type:
+        case 'MetaNetX':
+            for s in equation.split(' '):
+                # switch from reactants to products
+                if s == '=':
+                    is_product = True
+                # found stoichiometric factor
+                elif s.isnumeric():
+                    factor = float(s)
+                # skip
+                elif s == '+':
+                    continue
+                # found metabolite
+                else:
+                    # get information from MetaNetX
+                    metabolite, compartment = s.split('@')
+                    compartments.append(compartment)
+                    
+                    if is_product:
+                        products[metabolite] = factor
+                    else:
+                        reactants[metabolite] = factor
+                        
+        case 'BiGG':
+            factor = 1.0 # BiGG does not use factor 1 in the quations
+            for s in equation.split(' '):
+                # found factor
+                if s.replace('.','').isdigit():
+                    factor = float(s)
+                # switch from reactants to products
+                elif s == '-->' :
+                    is_product = True
+                    reversible = False
+                elif s == '<->':
+                    is_product = True
+                # skip
+                elif s == '+':
+                    continue
+                # found metabolite
+                else:
+                    compartments.append(s.rsplit('_',1)[1])
+                    if is_product:
+                        products[s] = factor
+                    else:
+                        reactants[s] = factor
+                    factor = 1.0
+              
+        case 'KEGG':
+            compartments = None
+            factor = 1.0
+            for s in equation.split(' '):
+                if s.isnumeric():
+                    factor = float(s)
+                elif s == '+':
+                    continue
+                elif s == '<=>': # @TODO are there more options?
+                    is_product = True
+                else:
+                    if is_product:
+                        products[s] = factor
+                    else:
+                        reactants[s] = factor
+                    factor = 1.0
+        
+        case 'BioCyc':
+            pass
+                  
+    return (reactants,products,compartments,reversible)
+        
+        
+ 
 
 # TODO
 #   extend the build function so, that all of them can take either the id or an equation
